folder structure update
